refactor(lib/validator): modularize checkInterfaceMemberDuplication
