+++ lib/productions/interface.js
@@ -5,6 +5,7 @@ import { Constant } from "./constant.js";
 import { IterableLike } from "./iterable.js";
 import { stringifier } from "./helpers.js";
 import { validationError } from "../error.js";
+import { checkInterfaceMemberDuplication } from "../validators/interface.js";
 
 /**
  * @param {import("../tokeniser").Tokeniser} tokeniser
@@ -52,5 +53,8 @@ for more information.`;
       yield validationError(this.source, this.tokens.name, this, message);
     }
     yield* super.validate(defs);
+    if (!this.partial) {
+      yield* checkInterfaceMemberDuplication(defs, this);
+    }
   }
 }
